Update to use the latest bakana v2 release.
This mostly involves rebasing (and renaming) the CollaboratorDbDataset to
derive from the ArtifactDB base class of the same name. As such, we can
delete a lot of the ArtifactDB loading code from this package.

We also updated some of the tests, mostly relating to serialization (as saving
analysis state is no longer supported).